Rename SceneRequestSupplementer to SupplementalRequestProvider
