+++ packages/scenes/src/querying/SceneQueryRunner.ts
@@ -34,7 +34,7 @@ import { writeSceneLog } from '../utils/writeSceneLog';
 import { VariableValueRecorder } from '../variables/VariableValueRecorder';
 import { emptyPanelData } from '../core/SceneDataNode';
 import { getClosest } from '../core/sceneGraph/utils';
-import { isRequestAdder, ProcessorFunc, SceneRequestSupplementer } from './SceneRequestAdder';
+import { isSupplementalRequestProvider, ProcessorFunc, SupplementalRequestProvider } from './SupplementalRequestProvider';
 import { passthroughProcessor, extraRequestProcessingOperator } from './extraRequestProcessingOperator';
 import { filterAnnotations } from './layers/annotations/filterAnnotations';
 import { getEnrichedDataRequest } from './getEnrichedDataRequest';
@@ -109,11 +109,11 @@ export class SceneQueryRunner extends SceneObjectBase<QueryRunnerState> implemen
 
   private _onActivate() {
     const timeRange = sceneGraph.getTimeRange(this);
-    const adders = this.getClosestRequestAdders();
-    for (const adder of adders) {
+    const providers = this.getClosestSupplementalRequestProviders();
+    for (const provider of providers) {
       this._subs.add(
-        adder.subscribeToState((n, p) => {
-          if (adder.shouldRerun(p, n)) {
+        provider.subscribeToState((n, p) => {
+          if (provider.shouldRerun(p, n)) {
             this.runQueries();
           }
         })
@@ -531,8 +531,8 @@ export class SceneQueryRunner extends SceneObjectBase<QueryRunnerState> implemen
     const primaryTimeRange = timeRange.state.value;
     let secondaryRequests: DataQueryRequest[] = [];
     let secondaryProcessors = new Map();
-    for (const adder of this.getClosestRequestAdders() ?? []) {
-      for (const { req, processor } of adder.getSupplementalRequests(request)) {
+    for (const provider of this.getClosestSupplementalRequestProviders() ?? []) {
+      for (const { req, processor } of provider.getSupplementalRequests(request)) {
         const requestId = getNextRequestId();
         secondaryRequests.push({ ...req, requestId })
         secondaryProcessors.set(requestId, processor ?? passthroughProcessor);
@@ -581,21 +581,21 @@ export class SceneQueryRunner extends SceneObjectBase<QueryRunnerState> implemen
   }
 
   /**
-   * Walk up the scene graph and find any request adders.
+   * Walk up the scene graph and find any SupplementalRequestProviders.
    *
-   * This will return a map from id to the closest adder for each id.
+   * This will return an array of the closest provider of each type.
    */
-  private getClosestRequestAdders(): SceneRequestSupplementer<any>[] {
+  private getClosestSupplementalRequestProviders(): Array<SupplementalRequestProvider<any>> {
     const found = new Map();
     if (!this.parent) {
       return [];
     }
     getClosest(this.parent, (s) => {
-      if (isRequestAdder(s) && !found.has(s.constructor.name)) {
+      if (isSupplementalRequestProvider(s) && !found.has(s.constructor.name)) {
         found.set(s.constructor.name, s);
       }
       s.forEachChild((child) => {
-        if (isRequestAdder(child) && !found.has(child.constructor.name)) {
+        if (isSupplementalRequestProvider(child) && !found.has(child.constructor.name)) {
           found.set(child.constructor.name, child);
         }
       });
